Added a copy of levenshtein_distance to ParticipantClass.py (called edit_distance) so it can be used in that Class.
Now initializing app_names and app_numbers with one empty row (each) so that, as app data is added, their ordinals are the same as their indexes in the DataFrames. (i.e. the 1st app on the screenshot has index 1, the 2nd has index 2, etc.) At the end, the leading empty row is removed before returning the dataframe of app names and numbers.

Will probably abandon the idea of 'subheading found' as a method for helping line up existing data with new data (when they're for the same person & day & category).