Modified how time text is filtered, so it doesn't have to re-merge an app name like '2 br' with a time like '34 min' into '2 hr 34 min', it will first check if the whole row is a misread time format.
Also, sometimes on the Google dashboard version of Android screenshots, there is a solid bar of irrelevant text above the first app, but when the cropped image is rescanned at a different bw threshold, this area can show faint signs of text. Now, the program searches for a solid bar of non-bg colour pixels above the first app, and if it finds one, it erases everything from that row and up.