Fixed error where detected language was French but the French entry in KEYWORDS_FOR_SHOW_SITES_YOU_VISIT was just an empty string, causing the code to try to calculate the log of 0. Added 'TODO FILL THIS IN' as the dictionary value, as well as for other empty dictionary values.
Also, fixed the warning when trying to convert a text date to a date object, where no year was provided (but one was requested). Current year is now provided; this will not change the year that gets stored for the screenshot, since the default '1900' year was already being replaced by the correct year.