Added 'android_version' to the output CSVs for screenshot data.
Added print statements after all screenshots have been analyzed to convey that the program is building the dataframes to output to CSVs.

Changed error message for 'unrelated data detected' to 'unreadable data format' -- this is more accurate, since some 'unrelated' images actually do contain relevant info, but the image is not in a format that the code has been programmed to read.