Improved the ability to read data from Android screenshots with a 'Search apps' (or 'Search categories') search bar above the app list. The 'Search apps' and 'Search categories' strings are now a new dictionary called SAMSUNG_SEARCH_APPS.
Also changed hash to be for the whole text_df from the initial text extraction from the screenshot. That way it will result in fewer false-positives.

Fixed iOS issue where sometimes 'erasing the bars and icons' would erase some of the left edge of the app text. Code now erases a percentage of the left edge, rather than all the pixels left of the textbox.