Replacing 'openal.com' with 'openai.com' in iOS get_app_names_and_numbers
Unified the replacing of misread app names (in iOS and Android) to be a single function in OCRScript_v3.py that both iOS and Android branches will call.

When app comparisons must be made (between new data and existing data) both sets of data are shown in full, to show not just the number of rows that will be compared but also all the rows of data extracted from both screenshots. This makes the comparison look more clear.