New method of detecting hours axis seems to be working just as well as the old method.
Also, found out why '00 06 12 18' was being replaced with '0 6 12 18' --- it has to do with the 'ensure_text_is_string' function in the extract_text_from_image function -- that function converted each word to an integer (if possible), then back to a string. The problem is now resolved, and the HOURS AXIS detection method no longer needs to include searches for '0' and '6', it can just include '00' and '06' (since that's how the hours row may appear on certain iOS screenshots).